[chore/#339] 패키징 분리
