--- data/user/src/main/java/com/hilingual/data/user/dto/reponse/UserInfoResponseDto.kt
@@ -1,33 +1,0 @@
-/*
- * Copyright 2025 The Hilingual Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.hilingual.data.user.dto.reponse
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class UserInfoResponseDto(
-    @SerialName("nickname")
-    val nickname: String,
-    @SerialName("profileImg")
-    val profileImg: String,
-    @SerialName("totalDiaries")
-    val totalDiaries: Int,
-    @SerialName("streak")
-    val streak: Int,
-    @SerialName("newAlarm")
-    val newAlarm: Boolean
-)
