+++ data/user/src/main/java/com/hilingual/data/user/dto/response/notification/NotificationDetailResponseDto.kt
@@ -1,4 +1,4 @@
-package com.hilingual.data.user.dto.reponse
+package com.hilingual.data.user.dto.response.notification
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
