+++ data/user/src/main/java/com/hilingual/data/user/dto/response/user/UserInfoResponseDto.kt
@@ -13,13 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hilingual.data.user.dto.reponse
+package com.hilingual.data.user.dto.response.user
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class NicknameResponseDto(
-    @SerialName("isAvailable")
-    val isAvailable: Boolean
+data class UserInfoResponseDto(
+    @SerialName("nickname")
+    val nickname: String,
+    @SerialName("profileImg")
+    val profileImg: String,
+    @SerialName("totalDiaries")
+    val totalDiaries: Int,
+    @SerialName("streak")
+    val streak: Int,
+    @SerialName("newAlarm")
+    val newAlarm: Boolean
 )
