+++ data/user/src/main/java/com/hilingual/data/user/model/user/NicknameValidationResult.kt
@@ -13,16 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hilingual.data.user.model
+package com.hilingual.data.user.model.user
 
-import com.hilingual.data.user.dto.request.UserProfileRequestDto
-
-data class UserProfileModel(
-    val profileImg: String,
-    val nickname: String
-)
-
-internal fun UserProfileModel.toDto() = UserProfileRequestDto(
-    profileImg = this.profileImg,
-    nickname = this.nickname
-)
+enum class NicknameValidationResult {
+    AVAILABLE,
+    DUPLICATE,
+    FORBIDDEN_WORD
+}
