+++ data/user/src/main/java/com/hilingual/data/user/repository/UserRepository.kt
@@ -15,12 +15,12 @@
  */
 package com.hilingual.data.user.repository
 
-import com.hilingual.data.user.model.NicknameValidationResult
-import com.hilingual.data.user.model.NotificationDetailModel
-import com.hilingual.data.user.model.NotificationModel
-import com.hilingual.data.user.model.NotificationSettingsModel
-import com.hilingual.data.user.model.UserInfoModel
-import com.hilingual.data.user.model.UserProfileModel
+import com.hilingual.data.user.model.user.NicknameValidationResult
+import com.hilingual.data.user.model.notification.NotificationDetailModel
+import com.hilingual.data.user.model.notification.NotificationModel
+import com.hilingual.data.user.model.notification.NotificationSettingsModel
+import com.hilingual.data.user.model.user.UserInfoModel
+import com.hilingual.data.user.model.user.UserProfileModel
 
 interface UserRepository {
     suspend fun getNicknameAvailability(
