+++ data/user/src/main/java/com/hilingual/data/user/repositoryimpl/UserRepositoryImpl.kt
@@ -18,14 +18,14 @@ package com.hilingual.data.user.repositoryimpl
 import com.hilingual.core.common.util.suspendRunCatching
 import com.hilingual.core.localstorage.UserInfoManager
 import com.hilingual.data.user.datasource.UserRemoteDataSource
-import com.hilingual.data.user.model.NicknameValidationResult
-import com.hilingual.data.user.model.NotificationDetailModel
-import com.hilingual.data.user.model.NotificationModel
-import com.hilingual.data.user.model.NotificationSettingsModel
-import com.hilingual.data.user.model.UserInfoModel
-import com.hilingual.data.user.model.UserProfileModel
-import com.hilingual.data.user.model.toDto
-import com.hilingual.data.user.model.toModel
+import com.hilingual.data.user.model.user.NicknameValidationResult
+import com.hilingual.data.user.model.notification.NotificationDetailModel
+import com.hilingual.data.user.model.notification.NotificationModel
+import com.hilingual.data.user.model.notification.NotificationSettingsModel
+import com.hilingual.data.user.model.user.UserInfoModel
+import com.hilingual.data.user.model.user.UserProfileModel
+import com.hilingual.data.user.model.user.toDto
+import com.hilingual.data.user.model.notification.toModel
 import com.hilingual.data.user.repository.UserRepository
 import javax.inject.Inject
