+++ data/user/src/main/java/com/hilingual/data/user/service/UserService.kt
@@ -16,12 +16,12 @@
 package com.hilingual.data.user.service
 
 import com.hilingual.core.network.BaseResponse
-import com.hilingual.data.user.dto.reponse.NicknameResponseDto
-import com.hilingual.data.user.dto.reponse.NotificationDetailResponseDto
-import com.hilingual.data.user.dto.reponse.NotificationResponseDto
-import com.hilingual.data.user.dto.reponse.NotificationSettingsResponseDto
-import com.hilingual.data.user.dto.reponse.UserInfoResponseDto
-import com.hilingual.data.user.dto.request.UserProfileRequestDto
+import com.hilingual.data.user.dto.response.user.NicknameResponseDto
+import com.hilingual.data.user.dto.response.notification.NotificationDetailResponseDto
+import com.hilingual.data.user.dto.response.notification.NotificationResponseDto
+import com.hilingual.data.user.dto.response.notification.NotificationSettingsResponseDto
+import com.hilingual.data.user.dto.response.user.UserInfoResponseDto
+import com.hilingual.data.user.dto.request.user.UserProfileRequestDto
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.PATCH
