+++ presentation/notification/src/main/java/com/hilingual/presentation/notification/main/model/NoticeNotificationItemUiModel.kt
@@ -1,7 +1,7 @@
 package com.hilingual.presentation.notification.main.model
 
 import androidx.compose.runtime.Stable
-import com.hilingual.data.user.model.NotificationModel
+import com.hilingual.data.user.model.notification.NotificationModel
 
 enum class NoticeCategoryType {
     SYSTEM,
